Changed call to update_readme
* Now it should be called though a main craft-ai command with option --update_readme
